Add binding for fluid_player_get_status()
Also refactor examples/player.mid to make it work on my system and to use
the new binding to poll the midi player status to know when the playback is
finished.  This allow to interrupt playback with Ctrl+C, which was not possible
with the previous implementation based on fluid_player_join().